refactor(classification): Move cycling parse utilities to domain directory
The diffs show moving parse-related files from utils/ to domain/cycling/
and removing several unused parser modules while updating import paths.
