+++ src/client/domain/cycling/parse/raceContent/@types/stages.js
